fix(firmware): audit P0/P1 — concurrence, sécurité actionneurs, stockage, réseau, auth
Audit complet de Servo_flute_ESP32 : les défauts corrigés ci-dessous étaient
tous exploitables en fonctionnement normal, pas seulement en théorie.

Concurrence (P0)
- EventQueue : `peek() -> lecture du pointeur -> dequeue()` n'était pas atomique.
  Entre les deux, une autre tâche pouvait vider la file, insérer un Note Off
  forcé (qui évince la tête sur file pleine), déplacer `_tail` ou déclencher un
  panic — d'où perte, réordonnancement ou exécution d'un événement obsolète.
  `tryPopDueEvent()` lit l'échéance, copie et retire le MÊME événement sous UN
  SEUL verrou ; `peekCopy()` rend une copie. Plus aucun pointeur interne n'est
  exposé, et `clear()` incrémente une époque qui abandonne une salve en cours.
- CommandQueue : les callbacks AsyncTCP et NimBLE pilotaient directement le bus
  I2C, les GPIO d'actionneurs et `cfg` pendant que loop() faisait de même. Ils
  postent désormais une commande appliquée par loop(), seul propriétaire du
  matériel. Le panic est un drapeau dédié (jamais perdu) qui annule les
  commandes antérieures. Les opérations web lourdes passent par un hand-off
  borné (bloquant côté HTTP, non bloquant côté WebSocket pour ne pas croiser le
  verrou d'AsyncWebSocket).

Configuration (P0)
- POST /api/config modifiait `cfg` champ par champ AVANT validation : loop()
  pouvait observer un état intermédiaire. Commit transactionnel ConfigCommit —
  candidat, normalisation, validation, décision de redémarrage, sauvegarde, puis
  activation atomique. Échec de validation ou de sauvegarde : rien ne bouge.
  Changement exigeant une ré-init : sauvegardé mais non activé jusqu'au reboot.
- validateAndNormalizeConfig() couvre les flottants (isfinite + bornes), tous
  les timings, seuils ADC, quotas, et ramène les chaînes libres à des ensembles
  fermés.

Sécurité actionneurs (P0)
- hardware_not_ready : les commandes web attaquaient getAirflowCtrl() /
  getPressureCtrl() / getFanCtrl() / le calibrateur directement et contournaient
  les gardes de noteOn()/setPWM() — pompes et ventilateur restaient pilotables
  après un échec PCA. Refus central dans InstrumentManager::applyCommand(),
  point d'application unique, plus réponse explicite côté web.
- LittleFS : `begin(true)` reformatait la partition au premier échec de montage,
  détruisant /config.json et les fichiers MIDI en silence. Montage sans
  formatage, mode recovery (OE haut, actionneurs interdits, diagnostic),
  formatage seulement sur POST /api/fs/format confirmé.

Correctifs P1/P2
- Vibrato : SIN_LUT est int8_t mais était lue par pgm_read_byte() (non signé) —
  la demi-période négative remontait entre 129 et 255, vibrato unipolaire.
- CC 120-127 ne passent plus par le limiteur de débit (124-127 pouvaient être
  jetés) ; CC2 sous le seuil de silence est toujours traité, au-dessus la
  dernière valeur est coalescée au lieu d'être perdue.
- Upload MIDI : état partagé par tout le serveur remplacé par un slot exclusif ;
  fichier temporaire unique ; validation complète AVANT remplacement du fichier
  existant (un upload invalide détruisait le fichier valide).
- VL53L0X : seule l'adresse 0x29 était sondée. TofSensor implémente la séquence
  complète (Model ID, data init, SPAD de référence, tuning ST, calibrations VHV
  et phase) et sépare présent / initialisé / mesure valide / périmée / en panne.
- Bascules AP<->STA centralisées : panic transport avant tout démontage, mDNS /
  rtpMIDI / DNS captif arrêtés et redémarrés exactement une fois.
- Clé hotspot : dérivée des 24 bits bas du MAC (diffusé en clair dans le BSSID).
  Remplacée par un secret aléatoire du RNG matériel, persistant en NVS.
- Authentification web : jeton de session sur toute route mutante et sur le
  WebSocket, mot de passe admin généré au premier boot, récupération par appui
  long sur BOOT au démarrage.
- JSON : toute chaîne libre passe par ArduinoJson.
- Diagnostics : état réel (PCA sondés, système de fichiers, capteur, transports,
  session actionneurs, commandes perdues) au lieu de "probe requires device".
- Qualité : code mort supprimé (valveServoDir, solenoidInterNoteMs, _lastCCTime,
  _pidLastError, alias legacy), comparaisons de rollover en int32_t explicite.

Tests
- tests/test_native/test_audit.cpp : 28 nouveaux tests comportementaux.
- tests/test_static_audit.py : 15 nouveaux verrous d'architecture, les 12 tests
  existants mis à jour pour la nouvelle architecture (aucun supprimé).
- Stub hôte corrigé : pgm_read_byte() rend désormais un octet NON signé comme
  sur ESP32, sinon le bug de vibrato restait invisible en test.

+++ Servo_flute_ESP32/SerialMidiHandler.cpp
@@ -75,9 +75,9 @@ void SerialMidiHandler::update() {
 
 void SerialMidiHandler::checkActiveSensingTimeout() {
   // Uniquement si la source s'est engagee via Active Sensing, et une seule fois
-  // par silence. (long)(...) rend la comparaison sure au rollover de millis().
+  // par silence. Le cast int32_t rend la comparaison sure au rollover de millis().
   if (!_activeSensing || _linkLost) return;
-  if ((long)(millis() - _lastByteMs) > MIDI_ACTIVE_SENSING_TIMEOUT_MS) {
+  if ((int32_t)(millis() - _lastByteMs) > MIDI_ACTIVE_SENSING_TIMEOUT_MS) {
     _linkLost = true;
     if (_instrument) _instrument->handleTransportLost();
     if (DEBUG) {
